Requests other than login post must not create attempt record.
The authenticator is called even if the request is not a post
nor the login URL.
In those cases, it will fail.
Since the wrong password was not entered,
the attempts should not be counted.

Record attempts only on identity not found failure.
Use constant of the interface.

In test, configure the login URL of the authenticator
and set URL to all requests.